+++ src/memory/mod.rs
@@ -2,4 +2,4 @@ pub mod bump_allocator;
 pub mod simple_memory;
 
 /// The number of bits per byte the memory system expects.
-pub const BITS_IN_BYTE: usize = 8;
+pub const BITS_IN_BYTE: u32 = 8;
